# src/main/resources/static/openapi/EducationAndWorkPlanAPI.yml
openapi: 3.0.0
info:
  title: Education and Work Plan API
  version: '1.7.5'
  description: Education and Work Plan API
  contact:
    name: Matt Wills
    email: matt.wills@digital.justice.gov.uk
servers:
  - url: 'http://localhost:3000'
#
# --------------------------------------------------------------------------------
#
paths:
...
  '/ciag/induction/{prisonNumber}':
    parameters:
      - name: prisonNumber
        description: The number of the Prisoner who may have relevant Induction data.
        schema:
          type: string
        example: 'A1234BC'
        in: path
        required: true
    post:
      deprecated: true
      summary: Creates an Induction for a Prisoner.
      description: |
        Creates an Induction for a Prisoner (if it does not already exist). This endpoint has been migrated from the
        `hmpps-ciag-careers-induction-api` and the data structure has been kept the same in order to prevent breaking
        the UI when it points to this endpoint.
      tags:
        - Legacy Inductions
      responses:
        '201':
          description: The Induction was created successfully.
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      operationId: create-ciag-prisoner-induction
      requestBody:
        $ref: '#/components/requestBodies/CreateCiagInduction'
    get:
      deprecated: true
      summary: Return a Prisoner's Induction.
      description: |
        Returns a Prisoner's Induction. Note that the response body (`CiagInduction`) matches the equivalent endpoint in the
        `hmpps-ciag-careers-induction-api`. This is to maintain compatibility for when the CIAG UI migrates to use this API instead.
      tags:
        - Legacy Inductions
      responses:
        '200':
          $ref: '#/components/responses/CiagInduction'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: get-ciag-prisoner-induction
    put:
      deprecated: true
      summary: Updates an Induction for a Prisoner.
      description: |
        Provides the ability to update a Prisoner's Induction. If a field (or "child" object) is provided in the request, it will
        completely overwrite the existing value. Conversely, if a field is not included in the request, then the existing value
        will remain untouched (including its updatedAt/updatedBy values). This endpoint has been migrated from the
        `hmpps-ciag-careers-induction-api` and the data structure has been kept the same in order to prevent breaking the CIAG
        UI when it points to this endpoint.
      tags:
        - Legacy Inductions
      responses:
        '204':
          description: CIAG Induction updated successfully.
        '400':
          $ref: '#/components/responses/400ErrorResponse'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: update-ciag-prisoner-induction
      requestBody:
        $ref: '#/components/requestBodies/UpdateCiagInduction'
...
  '/inductions/{prisonNumber}':
    parameters:
      - name: prisonNumber
        description: The number of the Prisoner who may have relevant Induction data.
        schema:
          type: string
        example: 'A1234BC'
        in: path
        required: true
    post:
      summary: Creates an Induction for a Prisoner.
      description: |
        Creates an Induction for a Prisoner (if it does not already exist).
      tags:
        - Inductions
      responses:
        '201':
          description: The Induction was created successfully.
        '400':
          $ref: '#/components/responses/400ErrorResponse'
      operationId: create-prisoner-induction
      requestBody:
        $ref: '#/components/requestBodies/CreateInduction'
    get:
      summary: Return a Prisoner's Induction.
      description: |
        Returns a Prisoner's Induction.
      tags:
        - Inductions
      responses:
        '200':
          $ref: '#/components/responses/Induction'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: get-prisoner-induction
    put:
      summary: Updates an Induction for a Prisoner.
      description: |
        Provides the ability to update a Prisoner's Induction. If a field (or "child" object) is provided in the request,
        it will completely overwrite the existing value. Conversely, if a field is not included in the request, then
        the existing value will remain untouched (including its updatedAt/updatedBy values).
      tags:
        - Inductions
      responses:
        '204':
          description: Induction updated successfully.
        '400':
          $ref: '#/components/responses/400ErrorResponse'
        '404':
          $ref: '#/components/responses/404ErrorResponse'
      operationId: update-prisoner-induction
      requestBody:
        $ref: '#/components/requestBodies/UpdateInduction'

#
# --------------------------------------------------------------------------------
#
components:
...
  schemas:
    ErrorResponse:
      title: ErrorResponse
      description: |
        A generic response body object describing errors in a web request, and can be used to communicate several
        different types of error condition such as (but not limited to) `400 BAD REQUEST`, `409 CONFLICT` etc.
      properties:
        status:
          description: The HTTP status code.
          type: integer
          example: 400
        errorCode:
          description: An optional application specific error code.
          type: string
        userMessage:
          description: An optional human readable description of the error.
          type: string
          example: 'No Action Plan found for prisoner [A1234BC]'
        developerMessage:
          description: An optional error message that may have more technical information for developers.
          type: string
        moreInfo:
          description: Optional more detailed information about the error.
          type: string
      required:
        - status
...
    CiagInductionResponse:
      title: CiagInductionResponse
      type: object
      description: A Prisoner's CIAG Induction containing information, such as their education history and future work interests.
      properties:
        reference:
          type: string
          format: uuid
          description: The Induction's unique reference.
          example: 814ade0a-a3b2-46a3-862f-79211ba13f7b
        offenderId:
          pattern: "^[A-Z]\\d{4}[A-Z]{2}$"
          type: string
          description: 'The ID of the Prisoner. AKA the prison number.'
        prisonId:
          type: string
          description: The ID of the Prison that that Prisoner is currently at.
        hopingToGetWork:
          $ref: '#/components/schemas/HopingToWork'
        desireToWork:
          type: boolean
          description: Whether the Prisoner wishes to work or not.
        reasonToNotGetWorkOther:
          type: string
          description: The reason that is given when the Prisoner does not want to work. This is mandatory when 'reasonToNotGetWork' is set to 'OTHER'.
        abilityToWork:
          uniqueItems: true
          type: array
          description: One or more factors affecting the Prisoner's ability to work.
          items:
            $ref: '#/components/schemas/AbilityToWorkFactor'
        abilityToWorkOther:
          type: string
          description: A specific factor affecting the Prisoner's ability to work. This is mandatory when 'abilityToWork' is set to 'OTHER'.
        reasonToNotGetWork:
          uniqueItems: true
          type: array
          description: One or more reasons the Prisoner has given not to get work after leaving Prison.
          items:
            $ref: '#/components/schemas/ReasonNotToWork'
        workExperience:
          "$ref": "#/components/schemas/PreviousWorkResponse"
        skillsAndInterests:
          "$ref": "#/components/schemas/SkillsAndInterestsResponse"
        qualificationsAndTraining:
          "$ref": "#/components/schemas/EducationAndQualificationResponse"
        inPrisonInterests:
          "$ref": "#/components/schemas/PrisonWorkAndEducationResponse"
        createdBy:
          type: string
          description: The DPS username of the person who created the Induction.
          example: 'asmith_gen'
        createdDateTime:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when the Induction was created.
          example: '2023-06-19T09:39:44Z'
        modifiedBy:
          type: string
          description: The DPS username of the person who last updated the Induction.
          example: 'asmith_gen'
        modifiedDateTime:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when the Goal was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
      required:
        - reference
        - offenderId
        - hopingToGetWork
        - createdBy
        - createdDateTime
        - modifiedBy
        - modifiedDateTime
    PreviousWorkResponse:
      title: PreviousWorkResponse
      description: A Prisoner's previous work experience. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: A unique reference for this Prisoner's previous work experience (not the database primary key).
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        hasWorkedBefore:
          type: boolean
        typeOfWorkExperience:
          uniqueItems: true
          type: array
          description: A list of the Prisoner's type of previous work experience.
          items:
            $ref: '#/components/schemas/WorkType'
        typeOfWorkExperienceOther:
          type: string
          description: A specific work experience type for the Prisoner. Mandatory when 'typeOfWorkExperience' includes 'OTHER'.
        workExperience:
          uniqueItems: true
          type: array
          description: A list of the Prisoner's previous work experience details.
          items:
            $ref: '#/components/schemas/WorkExperience'
        workInterests:
          $ref: '#/components/schemas/WorkInterestsResponse'
        modifiedBy:
          type: string
          description: The DPS username of the person who last updated this Prisoner's previous work.
          example: 'asmith_gen'
        modifiedDateTime:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when this Prisoner's previous work was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
      required:
        - hasWorkedBefore
        - modifiedBy
        - modifiedDateTime
    SkillsAndInterestsResponse:
      title: SkillsAndInterestsResponse
      description: Represents the personal skills and interests of a Prisoner. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: A unique reference for this Prisoner's skills and interests (not the database primary key).
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        skills:
          uniqueItems: true
          type: array
          description: One or more skills that the Prisoner feels they have.
          items:
            $ref: '#/components/schemas/PersonalSkillType'
        skillsOther:
          type: string
          description: A specific type of a skill that the Prisoner feels they have. Mandatory when 'skills' includes 'OTHER'.
        personalInterests:
          uniqueItems: true
          type: array
          description: One or more interests that the Prisoner feels they have.
          items:
            $ref: '#/components/schemas/PersonalInterestType'
        personalInterestsOther:
          type: string
          description: A specific type of a interest that the Prisoner feels they have. Mandatory when 'personalInterests' includes 'OTHER'.
        modifiedBy:
          type: string
          description: The DPS username of the person who last updated this Prisoner's skills and interests.
          example: 'asmith_gen'
        modifiedDateTime:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when this Prisoner's skills and interests was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
      required:
        - hasWorkedBefore
        - modifiedBy
        - modifiedDateTime
    EducationAndQualificationResponse:
      title: EducationAndQualificationResponse
      type: object
      description: A Prisoner's education and qualifications history. Migrated from hmpps-ciag-careers-induction-api.
      properties:
        id:
          type: string
          format: uuid
          description: A unique reference for this Prisoner's education and qualifications (not the database primary key).
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        educationLevel:
          $ref: '#/components/schemas/HighestEducationLevel'
        qualifications:
          uniqueItems: true
          type: array
          description: A list of the Prisoner's previous qualifications
          items:
            $ref: '#/components/schemas/AchievedQualification'
        additionalTraining:
          uniqueItems: true
          type: array
          description: Any additional training that the Prisoner has completed in the past.
          items:
            $ref: '#/components/schemas/TrainingType'
        additionalTrainingOther:
          type: string
          description: A specific type of training that does not fit the given 'additionalTraining' types. Mandatory when 'additionalTraining' includes 'OTHER'.
        modifiedBy:
          type: string
          description: The DPS username of the person who last updated this Prisoner's education and qualifications.
          example: 'asmith_gen'
        modifiedDateTime:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when this Prisoner's education and qualifications was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
      required:
        - modifiedBy
        - modifiedDateTime
    PrisonWorkAndEducationResponse:
      description: A Prisoner's in-prison work and education interests. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: A unique reference for this Prisoner's in-prison work and education interests (not the database primary key).
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        inPrisonWork:
          uniqueItems: true
          type: array
          description: A list of in-prison work that the Prisoner is interested in.
          items:
            $ref: '#/components/schemas/InPrisonWorkType'
        inPrisonWorkOther:
          type: string
          description: A specific type of in-prison work that does not fit the given 'inPrisonWork' types. Mandatory when 'inPrisonWork' includes 'OTHER'.
        inPrisonEducation:
          uniqueItems: true
          type: array
          description: Any potential in-prison education/training that the Prisoner is interested in.
          items:
            $ref: '#/components/schemas/InPrisonTrainingType'
        inPrisonEducationOther:
          type: string
          description: A specific type of in-prison education/training that does not fit the given 'inPrisonEducation' types. Mandatory when 'inPrisonEducation' includes 'OTHER'.
        modifiedBy:
          type: string
          description: The DPS username of the person who last updated this Prisoner's in-prison work and education interests.
          example: 'asmith_gen'
        modifiedDateTime:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when this Prisoner's in-prison work and education interests was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
      required:
        - modifiedBy
        - modifiedDateTime
...
    InductionResponse:
      title: InductionResponse
      type: object
      description: A Prisoner's Induction containing information, such as their education history and future work interests.
      properties:
        reference:
          type: string
          format: uuid
          description: A unique reference for this Induction.
          example: 814ade0a-a3b2-46a3-862f-79211ba13f7b
        workOnRelease:
          $ref: '#/components/schemas/WorkOnReleaseResponse'
        previousQualifications:
          $ref: '#/components/schemas/PreviousQualificationsResponse'
        previousTraining:
          $ref: '#/components/schemas/PreviousTrainingResponse'
        previousWorkExperiences:
          $ref: '#/components/schemas/PreviousWorkExperiencesResponse'
        inPrisonInterests:
          $ref: '#/components/schemas/InPrisonInterestsResponse'
        personalSkillsAndInterests:
          $ref: '#/components/schemas/PersonalSkillsAndInterestsResponse'
        futureWorkInterests:
          $ref: '#/components/schemas/FutureWorkInterestsResponse'
        createdBy:
          type: string
          description: The DPS username of the person who created the Induction.
          example: 'asmith_gen'
        createdByDisplayName:
          type: string
          description: The display name of the person who created the Induction.
          example: 'Alex Smith'
        createdAt:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when the Induction was created.
          example: '2023-06-19T09:39:44Z'
        createdAtPrison:
          type: string
          description: The identifier of the prison that the prisoner was resident at when the Induction was created.
          example: 'BXI'
        updatedBy:
          type: string
          description: The DPS username of the person who last updated the Induction.
          example: 'asmith_gen'
        updatedByDisplayName:
          type: string
          description: The display name of the person who last updated the Induction.
          example: 'Alex Smith'
        updatedAt:
          type: string
          format: date-time
          description: An ISO-8601 timestamp representing when the Induction was last updated. This will be the same as the created date if it has not yet been updated.
          example: '2023-06-19T09:39:44Z'
        updatedAtPrison:
          type: string
          description: The identifier of the prison that the prisoner was resident at when the Induction was updated.
          example: 'BXI'
      required:
        - reference
        - prisonNumber
        - workOnRelease
        - createdBy
        - createdByDisplayName
        - createdAt
        - createdAtPrison
        - updatedBy
        - updatedByDisplayName
        - updatedAt
        - updatedAtPrison
...
    CreateCiagInductionRequest:
      title: CreateCiagInductionRequest
      description: A request to create a CIAG Induction. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        prisonId:
          type: string
          description: The ID of the Prison that that Prisoner is currently at.
        hopingToGetWork:
          $ref: '#/components/schemas/HopingToWork'
        reasonToNotGetWorkOther:
          type: string
          description: The reason that is given when the Prisoner does not want to work. This is mandatory when 'reasonToNotGetWork' is set to 'OTHER'.
        abilityToWork:
          uniqueItems: true
          type: array
          description: One or more factors affecting the Prisoner's ability to work.
          items:
            $ref: '#/components/schemas/AbilityToWorkFactor'
        abilityToWorkOther:
          type: string
          description: A specific factor affecting the Prisoner's ability to work. This is mandatory when 'abilityToWork' is set to 'OTHER'.
        reasonToNotGetWork:
          uniqueItems: true
          type: array
          description: One or more reasons the Prisoner has given not to get work after leaving Prison.
          items:
            $ref: '#/components/schemas/ReasonNotToWork'
        workExperience:
          $ref: '#/components/schemas/CreatePreviousWorkRequest'
        skillsAndInterests:
          $ref: '#/components/schemas/CreateSkillsAndInterestsRequest'
        qualificationsAndTraining:
          $ref: '#/components/schemas/CreateEducationAndQualificationsRequest'
        inPrisonInterests:
          $ref: '#/components/schemas/CreatePrisonWorkAndEducationRequest'
      required:
        - hopingToGetWork
        - prisonId
    CreatePreviousWorkRequest:
      description: A request to persist the previous work experience of the Prisoner. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        hasWorkedBefore:
          type: boolean
        typeOfWorkExperience:
          uniqueItems: true
          type: array
          description: A list of the Prisoner's type of previous work experience.
          items:
            $ref: '#/components/schemas/WorkType'
        typeOfWorkExperienceOther:
          type: string
          description: A specific work experience type for the Prisoner. Mandatory when 'typeOfWorkExperience' includes 'OTHER'.
        workExperience:
          uniqueItems: true
          type: array
          description: A list of the Prisoner's previous work experience details.
          items:
            $ref: '#/components/schemas/WorkExperience'
        workInterests:
          $ref: '#/components/schemas/CreateWorkInterestsRequest'
      required:
        - hasWorkedBefore
...
    CreateSkillsAndInterestsRequest:
      description: A request to persist a Prisoner's personal skills and interests. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        skills:
          uniqueItems: true
          type: array
          description: One or more skills that the Prisoner feels they have.
          items:
            $ref: '#/components/schemas/PersonalSkillType'
        skillsOther:
          type: string
          description: A specific type of a skill that the Prisoner feels they have. Mandatory when 'skills' includes 'OTHER'.
        personalInterests:
          uniqueItems: true
          type: array
          description: One or more interests that the Prisoner feels they have.
          items:
            $ref: '#/components/schemas/PersonalInterestType'
        personalInterestsOther:
          type: string
          description: A specific type of a interest that the Prisoner feels they have. Mandatory when 'personalInterests' includes 'OTHER'.
    CreateEducationAndQualificationsRequest:
      description: A request to persist a Prisoner's education and qualifications. Migrated from hmpps-ciag-careers-induction-api.
        the inmate.
      type: object
      properties:
        educationLevel:
          $ref: '#/components/schemas/HighestEducationLevel'
        qualifications:
          uniqueItems: true
          type: array
          description: A list of the Prisoner's previous qualifications
          items:
            $ref: '#/components/schemas/AchievedQualification'
        additionalTraining:
          uniqueItems: true
          type: array
          description: Any additional training that the Prisoner has completed in the past.
          items:
            $ref: '#/components/schemas/TrainingType'
        additionalTrainingOther:
          type: string
          description: A specific type of training that does not fit the given 'additionalTraining' types. Mandatory when 'additionalTraining' includes 'OTHER'.
    CreatePrisonWorkAndEducationRequest:
      description: A request to persist a Prisoner's in-prison work and education interests. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        inPrisonWork:
          uniqueItems: true
          type: array
          description: A list of in-prison work that the Prisoner is interested in.
          items:
            $ref: '#/components/schemas/InPrisonWorkType'
        inPrisonWorkOther:
          type: string
          description: A specific type of in-prison work that does not fit the given 'inPrisonWork' types. Mandatory when 'inPrisonWork' includes 'OTHER'.
        inPrisonEducation:
          uniqueItems: true
          type: array
          description: Any potential in-prison education/training that the Prisoner is interested in.
          items:
            $ref: '#/components/schemas/InPrisonTrainingType'
        inPrisonEducationOther:
          type: string
          description: A specific type of in-prison education/training that does not fit the given 'inPrisonEducation' types. Mandatory when 'inPrisonEducation' includes 'OTHER'.
    UpdateCiagInductionRequest:
      title: UpdateCiagInductionRequest
      description: A request to update a CIAG Induction. Migrated from hmpps-ciag-careers-induction-api.
      type: object
      properties:
        reference:
          type: string
          format: uuid
          description: The Induction's unique reference. Currently optional (to avoid breaking the existing CIAG API), but will be mandatory in future (and will most likely become a path variable).
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        prisonId:
          type: string
          description: The ID of the Prison that that Prisoner is currently at.
        hopingToGetWork:
          $ref: '#/components/schemas/HopingToWork'
        reasonToNotGetWorkOther:
          type: string
          description: The reason that is given when the Prisoner does not want to work. This is mandatory when 'reasonToNotGetWork' is set to 'OTHER'.
        abilityToWork:
          uniqueItems: true
          type: array
          description: One or more factors affecting the Prisoner's ability to work.
          items:
            $ref: '#/components/schemas/AbilityToWorkFactor'
        abilityToWorkOther:
          type: string
          description: A specific factor affecting the Prisoner's ability to work. This is mandatory when 'abilityToWork' is set to 'OTHER'.
        reasonToNotGetWork:
          uniqueItems: true
          type: array
          description: One or more reasons the Prisoner has given not to get work after leaving Prison.
          items:
            $ref: '#/components/schemas/ReasonNotToWork'
        workExperience:
          $ref: '#/components/schemas/UpdatePreviousWorkRequest'
        skillsAndInterests:
          $ref: '#/components/schemas/UpdateSkillsAndInterestsRequest'
        qualificationsAndTraining:
          $ref: '#/components/schemas/UpdateEducationAndQualificationsRequest'
        inPrisonInterests:
          $ref: '#/components/schemas/UpdatePrisonWorkAndEducationRequest'
      required:
        - hopingToGetWork
        - prisonId
...
    HopingToWork:
      title: HopingToWork
      type: string
      description: Whether the Prisoner hopes to get work.
      enum:
        - "YES"
        - "NO"
        - "NOT_SURE"
    AbilityToWorkFactor:
      title: AbilityToWorkFactor
      description: Factors affecting a prisoner's ability to work.
      type: string
      enum:
        - CARING_RESPONSIBILITIES
        - LIMITED_BY_OFFENSE
        - HEALTH_ISSUES
        - NO_RIGHT_TO_WORK
        - OTHER
        - NONE
...
    ReasonNotToWork:
      title: ReasonNotToWork
      type: string
      description: Reasons the Prisoner has given not to get work after leaving Prison.
      enum:
        - LIMIT_THEIR_ABILITY
        - FULL_TIME_CARER
        - LACKS_CONFIDENCE_OR_MOTIVATION
        - HEALTH
        - RETIRED
        - NO_RIGHT_TO_WORK
        - NOT_SURE
        - OTHER
        - NO_REASON
...
    CreateInductionRequest:
      title: CreateInductionRequest
      description: A request to create an Induction.
      type: object
      properties:
        prisonId:
          type: string
          description: The identifier of the prison that the prisoner is currently resident at.
          example: BXI
        workOnRelease:
          $ref: '#/components/schemas/CreateWorkOnReleaseRequest'
        previousQualifications:
          $ref: '#/components/schemas/CreatePreviousQualificationsRequest'
        previousTraining:
          $ref: '#/components/schemas/CreatePreviousTrainingRequest'
        previousWorkExperiences:
          $ref: '#/components/schemas/CreatePreviousWorkExperiencesRequest'
        inPrisonInterests:
          $ref: '#/components/schemas/CreateInPrisonInterestsRequest'
        personalSkillsAndInterests:
          $ref: '#/components/schemas/CreatePersonalSkillsAndInterestsRequest'
        futureWorkInterests:
          $ref: '#/components/schemas/CreateFutureWorkInterestsRequest'
      required:
        - prisonId
        - workOnRelease
...
    UpdateInductionRequest:
      title: UpdateInductionRequest
      description: A request to update a Prisoner's Induction.
      type: object
      properties:
        reference:
          type: string
          format: uuid
          description: The Induction's unique reference.
          example: c88a6c48-97e2-4c04-93b5-98619966447b
        workOnRelease:
          $ref: '#/components/schemas/UpdateWorkOnReleaseRequest'
        previousQualifications:
          $ref: '#/components/schemas/UpdatePreviousQualificationsRequest'
        previousTraining:
          $ref: '#/components/schemas/UpdatePreviousTrainingRequest'
        previousWorkExperiences:
          $ref: '#/components/schemas/UpdatePreviousWorkExperiencesRequest'
        inPrisonInterests:
          $ref: '#/components/schemas/UpdateInPrisonInterestsRequest'
        personalSkillsAndInterests:
          $ref: '#/components/schemas/UpdatePersonalSkillsAndInterestsRequest'
        futureWorkInterests:
          $ref: '#/components/schemas/UpdateFutureWorkInterestsRequest'
      required:
        - reference
...
  responses:
    400ErrorResponse:
      description: Error response for a HTTP 400 describing fields in the request payload that are in error.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            userMessage: 'Error on field goal.targetCompletionDate: rejected value [2023-06-01] cannot be in the past'
    404ErrorResponse:
      description: |
        Error response for a HTTP 404 describing the problem with the request.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            userMessage: 'The prisoner with prisonNumber=[A1234BC] not found'
...
    CiagInduction:
      description: Response body containing a Prisoner's CIAG Induction.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CiagInductionResponse'
...
    Induction:
      description: Response body containing a Prisoner's Induction.
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
        Access-Control-Allow-Methods:
          schema:
            type: string
        Access-Control-Allow-Headers:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InductionResponse'

  #
  # Request Body Definitions
  # --------------------------------------------------------------------------------
  requestBodies:
...
    CreateCiagInduction:
      description: Request body to create a Prisoner's Induction.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateCiagInductionRequest'
    UpdateCiagInduction:
      description: Request body to update a Prisoner's Induction.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateCiagInductionRequest'
...
    CreateInduction:
      description: Request body to create a Prisoner's Induction.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreateInductionRequest'
    UpdateInduction:
      description: Request body to update a Prisoner's Induction.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateInductionRequest'